Job Worker Library
Changes to support StartJob, StopJob and StreamJobOutput and
GetJobStatus library functions. The code has comprehensive logging
throughout for easier debugging.

Contains one integration test to cover most of the happy path logic

The test:
Starts a job using the JobManager.
Checks if the job was created successfully.
Waits for the job to complete.
Verifies the job's final status and exit code.
Checks if a log file was created for the job and if it contains data.
Finally, it streams the job's output.

Running the test:
jobworker$ make integration-test

The test should be run with "sudo" as the code needs to create cgroups
and namespaces.

More tests to come.